+++ components/catalog/CatalogPublicPage.tsx
@@ -1,73 +1,17 @@
 'use client';
 
 import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardFooter } from '@/components/ui/card';
 import { Separator } from '@/components/ui/separator';
-import { Link as LinkIcon, Clipboard, ChevronDown, ChevronUp, EyeIcon, Search, FilterIcon, Loader2, Mars, Venus, CircleHelp } from 'lucide-react';
-import Image from 'next/image';
+import { Link as LinkIcon, Clipboard, ChevronDown, ChevronUp, Search, FilterIcon, Loader2} from 'lucide-react';
 import { toast } from 'sonner';
 import { useState } from 'react';
 import { APP_URL } from '@/lib/constants/app';
 import { useQuery } from '@tanstack/react-query';
 import { getCatalogEntriesByProfileName } from '@/app/api/catalog';
 import { EnrichedCatalogEntry } from '@/lib/types/catalog';
-import { extractLastTwoDigitsOfYear } from '@/lib/utils';
+import ReptileCard from './ReptileCard';
 
 
-
-interface ReptileCardProps {
-  entry: EnrichedCatalogEntry;
-}
-
-function ReptileCard({ entry }: ReptileCardProps) {
-  const reptile = entry.reptiles;
-  const imageUrl = entry.catalog_images?.[0]?.image_url;
-
-  return (
-    <Card className="overflow-hidden transition-all hover:shadow-md py-0 gap-0">
-      <div className="relative aspect-square bg-muted">
-        {imageUrl ? (
-          <Image
-            src={imageUrl}
-            alt={reptile.name}
-            fill
-            className="object-cover"
-            sizes="(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw"
-          />
-        ) : (
-          <div className="flex items-center justify-center h-full w-full">
-            <EyeIcon className="h-8 w-8 text-muted-foreground/50" />
-          </div>
-        )}
-      </div>
-      <CardContent className="p-4">
-          <h3 className="text-xs md:text-[0.9rem] 3xl:text-base font-medium min-h-[30px] sm:min-h-[40px] tracking-wide">{reptile.name}</h3>       
-          <p className="text-xs sm:text-sm text-muted-foreground truncate">
-            {reptile.morph.name}
-          </p>
-          <div className="flex items-center gap-1.5">
-                 <div>
-                  {reptile.sex === 'male' ? (
-                        <Mars className="h-3.5 w-3.5 text-blue-400 mt-0.5 shrink-0" />
-                      ) : reptile.sex === 'female' ? (
-                        <Venus className="h-3.5 w-3.5 text-red-500 mt-0.5 shrink-0" />
-                      ) : (
-                        <CircleHelp className="h-3.5 w-3.5 text-muted-foreground mt-0.5 shrink-0" />
-                      )}
-                </div>
-                <p className="text-xs md:text-sm text-muted-foreground truncate">
-                  {extractLastTwoDigitsOfYear(reptile.hatch_date)}
-                </p>
-          </div>
-      </CardContent>
-      <CardFooter className="p-4 pt-0 text-sm text-muted-foreground">
-        <div className="w-full truncate">
-       
-        </div>
-      </CardFooter>
-    </Card>
-  );
-}
 
 interface CatalogClientPageProps {
   profileName: string;
